🎨 Refactor: Improve Currency Input UX
Refactor the currency input to use a standard select element instead of the custom `CurrencyInput` component, as the latter was unused.
This change enhances the user experience by simplifying the input and aligning it with standard UI conventions.
The `CurrencyInput` component has been marked as TODO and can be revisited later.