:rocket: Adds more investment data
Adds a large number of new investment options to the application's database.  This expands the range of investment choices available to users.  The seed script has been removed as the new data is directly included in the `placeholder-data.js` file.  A unique constraint is added to the database schema to prevent duplicate investments for a given user.  The README is updated to clarify the project's origins and give appropriate credit.